% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_from_gis_us.R
\name{adj_from_gis_us}
\alias{adj_from_gis_us}
\title{Generate an adjacency matrix corresponding to the "United States" graph}
\usage{
adj_from_gis_us()
}
\value{
An adjacency matrix
}
\description{
Generate an adjacency matrix corresponding to the "United States" graph
}
