% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_from_graph.R
\name{adj_from_graph}
\alias{adj_from_graph}
\title{Generate an adjacency matrix from a graph}
\usage{
adj_from_graph(graph)
}
\arguments{
\item{graph}{Graph, represented as a data frame with two columns ('n1' and
'n2'). Each row represents an edge between two nodes. Values in the
data frame represent the index of the node}
}
\value{
An adjacency matrix
}
\description{
Generate an adjacency matrix from a graph
}
