% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cov_car.R
\name{generate_cov_car}
\alias{generate_cov_car}
\title{Create a covariange matrix (CAR model) from an adjacency matrix}
\usage{
generate_cov_car(adj_mtx, rho)
}
\arguments{
\item{adj_mtx}{Adjacency matrix}

\item{rho}{Spatial correlation parameter}
}
\value{
A list containing the following: \cr
    * `mtx_cov`: Covariance matrix corresponding to the given adjacency
        matrix
    * `mtx_D`: D matrix
    * `n_i`: Vector containing number of neighbors for region i
    * `neighbors`: List containing neighbor sets for each region
    * `dim`: Dimension (number of regions)
}
\description{
Create a covariange matrix (CAR model) from an adjacency matrix
}
