% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_graph_grid.R
\name{generate_graph_grid}
\alias{generate_graph_grid}
\title{Generate a graph corresponding to the "grid" graph}
\usage{
generate_graph_grid()
}
\value{
A data frame representing the edges of a graph
}
\description{
Numbering starts on the top left (#1) and goes to the top right
    (#10), then continues on the next row (#11), etc.
}
