% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_graph_path.R
\name{generate_graph_path}
\alias{generate_graph_path}
\title{Generate a graph corresponding to the "path" graph}
\usage{
generate_graph_path()
}
\value{
A data frame representing the edges of a graph
}
\description{
Numbering starts on the left (#1) and goes to the right (#100)
}
