% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_matern_cov.R
\name{generate_matern_cov}
\alias{generate_matern_cov}
\title{Generate a matern covariance matrix from a graph}
\usage{
generate_matern_cov(type, phi)
}
\arguments{
\item{type}{Type of dataset (either "path", "grid", or "US")}

\item{phi}{Covariance matrix parameter}
}
\value{
A covariance matrix
}
\description{
Generate a matern covariance matrix from a graph
}
