% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mtx.R
\name{generate_mtx}
\alias{generate_mtx}
\title{Generate CAR/DAGAR model matrix components}
\usage{
generate_mtx(model, adj_mtx, rho = 1)
}
\arguments{
\item{model}{One of: "iCAR", "CAR", "Scaled iCAR", "DAGAR", "DAGAR_OF"}

\item{adj_mtx}{An adjacency matrix}
}
\value{
A list containing the following: \cr
    * `D`: Distance matrix for CAR-type matrices \cr
    * `n_i`: Vector of number of neighbors, indexed by region i \cr
    * `Q`: Precision matrix (with tau_W param set to 1) \cr
    # `neighbors`: A list of neighbors, indexed by region i
}
\description{
Generate CAR/DAGAR model matrix components
}
