% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate one dataset !!!!! TO DO}
\usage{
generate_dataset(type, rho)
}
\arguments{
\item{type}{Type of dataset (either "path", "grid", or "US")}

\item{rho}{Spatial correlation parameter}
}
\value{
A list containing the following: \cr
    * `params`: A list of the parameters supplied in the function call \cr
    * `reg_data`: A data frame of data points corresponding to a single
        region \cr
    * `adj_mtx`: The adjacency matrix \cr
    * `x`: X \cr
    * `x`: X \cr
    * `x`: X \cr
    * `x`: X \cr
    * `x`: X
}
\description{
Generate one dataset !!!!! TO DO
}
